Rename storage to db
